+++ example/android/app/src/main/kotlin/com/example/multi_video_player_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.multi_video_player_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}