+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## [0.0.1]
 
 Multi video player providing functionality to play multiple preloaded videos.
+
+## [0.0.2]
+
+* Added parameter showControlsOverlay
+* Added parameter showVideoProgressIndicator